requirements.txt
asgiref==3.4.1
Django==4.0
django-phonenumber-field==6.0.0
gunicorn==20.1.0
phonenumbers==8.12.40
psycopg2==2.9.3
python-dotenv==0.19.2
sqlparse==0.4.2
djangorestframework==3.11.0
pytz==2021.3